iOS-155: Remove 2 steps and tiers code
