+++ RevenueCatUI/Localizations.swift
@@ -11,16 +11,7 @@ import SwiftUI
 
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 public func setupPaywall(
-    paywallFirstStepBundle: Bundle,
     paywallSecondStepBundle: Bundle,
-    linTemplate5Bundle: Bundle,
-    timelineBundle: Bundle,
-    testimonialsBundle: Bundle
 ) {
-    LinTemplatesResources.linTemplate5Step1Bundle = paywallFirstStepBundle
     LinTemplatesResources.linTemplate5Step2Bundle = paywallSecondStepBundle
-    LinTemplatesResources.linTemplate5Bundle = linTemplate5Bundle
-    CTAFooterMessageProvider.bundle = paywallSecondStepBundle
-    TimelineView.bundle = timelineBundle
-    TestimonialsView.bundle = testimonialsBundle
 }
