+++ RevenueCatUI/Templates/LinTemplates/Helpers/LinTemplatesResources.swift
@@ -11,7 +11,5 @@ import SwiftUI
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 struct LinTemplatesResources {
     
-    static var linTemplate5Step1Bundle = Foundation.Bundle.module
     static var linTemplate5Step2Bundle = Foundation.Bundle.module
-    static var linTemplate5Bundle = Foundation.Bundle.module
 }
